feat: enhance comment upvote functionality with designId and add routes

+++ backend/src/services/desginCommentVote.service.js
@@ -2,36 +2,40 @@ import { DesignCommentVote } from "../models/design_comment_votes.model.js"
 import { DesignComment } from "../models/design_comments.model.js"
 import { ApiError } from "../utils/ApiError.js"
 
-const createCommentUpVote=async({commentId, userId})=>{
+const createCommentUpVote=async({commentId,designId, userId})=>{
     const comment= await DesignComment.findById(commentId)
     if(!comment){
         throw new ApiError(404, "Comment not found")
     }
+    if (comment.designId.toString() !== designId.toString()) {
+        throw new ApiError(404, "Comment not found");
+    }
     const existingUpVote= await DesignCommentVote.findOne({commentId:commentId,userId:userId})
     if(existingUpVote){
-        throw new ApiError(400, "cant upvote the comment again")
+        throw new ApiError(409, "cant upvote the comment again")
     }
     await DesignCommentVote.create({
         commentId,
         userId
-    })
-    const designId=comment.designId
-    await DesignComment.findByIdAndUpdate(designId, { $inc: { upvoteCount: 1 } });
+    })    
+    await DesignComment.findByIdAndUpdate(commentId, { $inc: { upvoteCount: 1 } });
     return true
 }
 
-const deleteCommentUpVote=async({commentId,userId})=>{
+const deleteCommentUpVote=async({commentId,designId,userId})=>{
     const comment= await DesignComment.findById(commentId)
     if(!comment){
         throw new ApiError(404, "Comment not found")
     }
+    if (comment.designId.toString() !== designId.toString()) {
+        throw new ApiError(404, "Comment not found");
+    }
     const existingUpVote= await DesignCommentVote.findOne({commentId:commentId,userId:userId})
     if(!existingUpVote){
-        throw new ApiError(400, "cant downvote the comment again")
+        throw new ApiError(404, "cant downvote the comment again")
     }
     await DesignCommentVote.findOneAndDelete({commentId,userId})
-    const designId=comment.designId
-    await DesignComment.findByIdAndUpdate(designId, { $inc: { upvoteCount: -1 } });
+    await DesignComment.findByIdAndUpdate(commentId, { $inc: { upvoteCount: -1 } })
     return true
 }
